Agrego 5% de ataque en territorio enemigo

+++ src/main/java/fiuba/algo3/algoChess/modelo/celda/Posicionable.java
@@ -14,5 +14,5 @@ public interface Posicionable {
 
     void addObserver(Observador observador);
 
-    int obtenerVida();
+    double obtenerVida();
 }
